backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
PyPDF2==3.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.0
lxml==4.9.3
